Fix for notifications not being handled correctly if they are sent with an interval of time too short between them.
Npgsql notification handling depends on Socket.Poll to check if there is
data to be read. Npgsql uses a BufferedStream which defeat the purpose
of Socket.Poll because it buffers all data and so Poll can't check if
there is more data to be read.

We "fix" that by using a non buffered stream when the notification handling is in place.

Note that this "fix" won't work on SSL connections. SSL Streams do their own buffering.

This all should be changed after we move to an async socket approach where we won't need to use Poll.